Rename "config/config_main.h" to "config.h"
