+++ src/config/config_legacy_settings.cpp
@@ -18,7 +18,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "config/config_main.h"
+#include "config.h"
 #include "globals.h"
 
 #include <IniProcessor/ini_processing.h>
